fix msg format bugs

+++ siemens_plc/scripts/siemens_plc_interface.py
@@ -30,7 +30,7 @@ def laser_config_param_callback(config, level):
 def check_laser_switch_callback(msg, args):
     # get the laser_switch msgs and make laser_config array
     # TODO: change the laser config type based on your needs(int32 is now right).
-    args[1].data = [laser_config_1, laser_config_2, laser_config_3, msg.laser_switch.data]
+    args[1].data = [np.int32(laser_config_1), np.int32(laser_config_2), np.int32(laser_config_3), np.int32(msg.laser_switch.data)]
     
     # publish the laser config parameters
     args[0].publish(args[1])
